cleans up and unifies category definitions
