+++ app/pages/HomePage/HomePage.tsx
@@ -59,51 +59,3 @@ export const HomePage = () => {
     </>
   );
 };
-
-/* TODO: Remove when new categories are created
- other components are dependent on this list */
-
-export const coreCategories = [
-  {
-    algoliaCategoryName: "Arts, Culture & Identity",
-    name: "Arts, Culture & Identityasdf",
-    icon: "food",
-    categorySlug: "arts-culture-identity",
-  },
-  {
-    algoliaCategoryName: "Children's Care",
-    name: "Children's Care",
-    icon: "food",
-    categorySlug: "childrens-care",
-  },
-  {
-    algoliaCategoryName: "Education",
-    name: "Education",
-    icon: "hospital",
-    categorySlug: "education",
-  },
-  {
-    algoliaCategoryName: "Family Support",
-    name: "Family Support",
-    icon: "shower",
-    categorySlug: "family-support",
-  },
-  {
-    algoliaCategoryName: "Health & Wellness",
-    name: "Health & Wellness",
-    icon: "bed",
-    categorySlug: "health-wellness",
-  },
-  {
-    algoliaCategoryName: "Sports & Recreation",
-    name: "Sports & Recreation",
-    icon: "longterm-housing",
-    categorySlug: "sports-recreation",
-  },
-  {
-    algoliaCategoryName: "Youth Workforce & Life Skills",
-    name: "Youth Workforce & Life Skills",
-    icon: "housing-heart",
-    categorySlug: "youth-workforce-life-skills",
-  },
-];
